v0.1.0 - Routing, Background, Theme, Languages
Cleaned up the code and changed some general styles.

Added a navigation bar with links to future pages.
There is also a theme toggle button and a language selector.
The theme and language are automatically detected based on the user's settings and can be changed at any time.

Added page routing.
The navigation link for the current page and its group are highlighted.

Added a footer.
Clicking the logo scrolls the page to the top.

Added a gradient background that changes on page scroll.

Implemented appropriate tests to verify functionality.