v0.2.0 - Navigation, Converters, Cypress Setup
Updated the navigation section to automatically generate links, groups and buttons from a JSON file and assign corresponding icons.
Content for converter pages is also generated from JSON files.
If a page or its contents are not found, the user is redirected to the 404 Not Found page.

Set up the Cypress E2E testing environment.
Implemented E2E routing tests.

Modified the mobile navigation popup to close when the "home" link is clicked, similar to the other links.

Fixed the app favicon.