v0.6.1 - Routing & Translation Fix
Fixed going back to the previous page from the "Not Found" page.
The user is now correctly redirected to the last valid page instead of being stuck in a "Not Found" page loop caused by going back to the page that does not exist, which redirects to the "Not Found" page.

Changed the paths to the converter files (JSON) to be absolute so they work correctly in a production build.