v0.6.2 - Loading Data Screen
Added a "Loading" screen to the Converter page.
It shows a loading icon when the data takes too long to load.
The icon doesn't show up immediately, but fades in after half a second to prevent flashing on page load
It also made the page switching transition look smoother, as the footer is now not placed at the bottom of the page and not jumping to the top.

Fixed a bug where the placeholder was not centered when there was no content in the operation field.